# docs/api-documentation.yaml
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1.0/auth/register": {
      "post": {
        "description": "Creates a new user account with the provided details",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1.0/auth/login": {
      "post": {
        "description": "Login with the provided details (JWT set as HTTPOnly cookie)",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User login credentials",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login using email and password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1.0/auth/me": {
      "get": {
        "description": "Returns profile details of the currently authenticated user from the JWT token.",
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile successfully fetched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get current user information",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1.0/auth/logout": {
      "post": {
        "description": "Clears authentication cookies (access_token and refresh_token).",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Logout user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1.0/auth/check-email": {
      "post": {
        "description": "Verifies whether the given email is already registered in the system.",
        "operationId": "AuthController_checkEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Email to be checked",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email is available for registration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Email already exists in the system",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Check if an email already exists",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1.0/auth/verify-otp": {
      "post": {
        "description": "Verifies the provided OTP for the given email address.",
        "operationId": "AuthController_verifyEmailOtp",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify the email OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1.0/auth/verify-recaptcha": {
      "post": {
        "description": "Endpoint to verify a user is human before allowing other actions.",
        "operationId": "AuthController_verifyRecaptcha",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecaptchaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Human verification successful."
          },
          "400": {
            "description": "reCAPTCHA verification failed."
          }
        },
        "summary": "Verifies a Google reCAPTCHA token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1.0/auth/forgotPassword": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset link successfully sent to the provided email",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "message": "Check your email for password reset instructions"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid email format"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Request a password reset link",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1.0/auth/verifyResetToken": {
      "get": {
        "operationId": "AuthController_verifyResetToken",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "example": "1",
              "type": "number"
            }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "example": "reset-token-from-email",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "message": "Token is valid",
                    "data": {
                      "valid": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token invalid or expired"
          }
        },
        "summary": "Verify if a reset token is valid",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1.0/auth/resetPassword": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password successfully reset",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "success",
                    "message": "Password has been reset successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid password format"
          },
          "401": {
            "description": "Token invalid or expired"
          }
        },
        "summary": "Reset password using valid token",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1.0/auth/test": {
      "get": {
        "description": "A protected test endpoint to verify JWT authentication.",
        "operationId": "AuthController_test",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful test",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Test endpoint",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1.0/auth/update-email": {
      "patch": {
        "description": "Updates the email address of the currently authenticated user.",
        "operationId": "AuthController_updateEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Email already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update user email",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1.0/auth/update-username": {
      "patch": {
        "description": "Updates the username of the currently authenticated user.",
        "operationId": "AuthController_updateUsername",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsernameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Username updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Username already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update username",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hankers",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name for the user",
            "example": "Mohaned Albaz",
            "minLength": 3,
            "maxLength": 30
          },
          "email": {
            "type": "string",
            "description": "The email address of the user",
            "example": "mohmaedalbaz@gmail.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "The password for the user account (must include uppercase, lowercase, number, and special character)",
            "example": "Password123!",
            "minLength": 8,
            "maxLength": 50,
            "format": "password"
          },
          "birth_date": {
            "type": "string",
            "description": "The birth date of the user",
            "example": "2004-01-01",
            "format": "date"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "birth_date"
        ]
      },
...
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Account created successfully."
          },
          "data": {
            "$ref": "#/components/schemas/RegisterDataResponseDto"
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "error",
              "fail"
            ],
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Invalid input data"
          },
          "error": {
            "type": "object",
            "nullable": true,
            "example": "Bad Request",
            "description": "Optional error details or the type of error"
          }
        },
        "required": [
          "status",
          "message",
          "error"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "mohamedalbaz@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "Test1234!",
            "description": "User password (min 8 characters)"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Logged in successfully"
          },
          "data": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "status",
          "message",
          "data"
        ]
      },
...
      "CheckEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "mohamedalbaz@gmail.com",
            "description": "The email address to check for existence"
          }
        },
        "required": [
          "email"
        ]
      },
...
    }
  }
}